Fases Finales

+++ ProyectoFinal_DuranDaniel/Controllers/CursoesController.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using ProyectoFinal_DuranDaniel.Data;
@@ -19,88 +15,98 @@ namespace ProyectoFinal_DuranDaniel.Controllers
             _context = context;
         }
 
-        // GET: Cursoes
+        // GET: /Cursoes/Index
         public async Task<IActionResult> Index()
         {
-            var appDbContext = _context.Cursos.Include(c => c.Carrera).Include(c => c.Docente);
-            return View(await appDbContext.ToListAsync());
+            var rol = HttpContext.Session.GetString("UsuarioRol");
+            if (string.IsNullOrEmpty(rol))
+                return RedirectToAction("Login", "Auth");
+
+            var cursos = await _context.Cursos
+                .Include(c => c.Carrera)
+                .Include(c => c.Docente)
+                .ToListAsync();
+
+            return View(cursos);
         }
 
-        // GET: Cursoes/Details/5
+        // GET: /Cursoes/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            var rol = HttpContext.Session.GetString("UsuarioRol");
+            if (string.IsNullOrEmpty(rol))
+                return RedirectToAction("Login", "Auth");
+
+            if (id == null) return NotFound();
 
             var curso = await _context.Cursos
                 .Include(c => c.Carrera)
                 .Include(c => c.Docente)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (curso == null)
-            {
-                return NotFound();
-            }
 
+            if (curso == null) return NotFound();
             return View(curso);
         }
 
-        // GET: Cursoes/Create
+        // GET: /Cursoes/Create
         public IActionResult Create()
         {
-            ViewData["CarreraId"] = new SelectList(_context.Carreras, "Id", "Id");
-            ViewData["DocenteId"] = new SelectList(_context.Usuarios, "Id", "Id");
+            if (HttpContext.Session.GetString("UsuarioRol") != "Administrador")
+                return RedirectToAction("Index");
+
+            ViewData["CarreraId"] = new SelectList(_context.Carreras, "Id", "Nombre");
+            ViewData["DocenteId"] = new SelectList(
+                _context.Usuarios.Where(u => u.RolId == 2), "Id", "Nombre");
             return View();
         }
 
-        // POST: Cursoes/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: /Cursoes/Create
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Nombre,Descripcion,CarreraId,DocenteId")] Curso curso)
         {
+            if (HttpContext.Session.GetString("UsuarioRol") != "Administrador")
+                return RedirectToAction("Index");
+
             if (ModelState.IsValid)
             {
                 _context.Add(curso);
                 await _context.SaveChangesAsync();
+                TempData["Mensaje"] = "Curso agregado correctamente.";
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["CarreraId"] = new SelectList(_context.Carreras, "Id", "Id", curso.CarreraId);
-            ViewData["DocenteId"] = new SelectList(_context.Usuarios, "Id", "Id", curso.DocenteId);
+
+            ViewData["CarreraId"] = new SelectList(_context.Carreras, "Id", "Nombre", curso.CarreraId);
+            ViewData["DocenteId"] = new SelectList(
+                _context.Usuarios.Where(u => u.RolId == 2), "Id", "Nombre", curso.DocenteId);
             return View(curso);
         }
 
-        // GET: Cursoes/Edit/5
+        // GET: /Cursoes/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (HttpContext.Session.GetString("UsuarioRol") != "Administrador")
+                return RedirectToAction("Index");
 
+            if (id == null) return NotFound();
             var curso = await _context.Cursos.FindAsync(id);
-            if (curso == null)
-            {
-                return NotFound();
-            }
-            ViewData["CarreraId"] = new SelectList(_context.Carreras, "Id", "Id", curso.CarreraId);
-            ViewData["DocenteId"] = new SelectList(_context.Usuarios, "Id", "Id", curso.DocenteId);
+            if (curso == null) return NotFound();
+
+            ViewData["CarreraId"] = new SelectList(_context.Carreras, "Id", "Nombre", curso.CarreraId);
+            ViewData["DocenteId"] = new SelectList(
+                _context.Usuarios.Where(u => u.RolId == 2), "Id", "Nombre", curso.DocenteId);
             return View(curso);
         }
 
-        // POST: Cursoes/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
+        // POST: /Cursoes/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Nombre,Descripcion,CarreraId,DocenteId")] Curso curso)
         {
-            if (id != curso.Id)
-            {
-                return NotFound();
-            }
+            if (HttpContext.Session.GetString("UsuarioRol") != "Administrador")
+                return RedirectToAction("Index");
+
+            if (id != curso.Id) return NotFound();
 
             if (ModelState.IsValid)
             {
@@ -108,57 +114,56 @@ namespace ProyectoFinal_DuranDaniel.Controllers
                 {
                     _context.Update(curso);
                     await _context.SaveChangesAsync();
+                    TempData["Mensaje"] = "Curso actualizado correctamente.";
                 }
                 catch (DbUpdateConcurrencyException)
                 {
                     if (!CursoExists(curso.Id))
-                    {
                         return NotFound();
-                    }
                     else
-                    {
                         throw;
-                    }
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["CarreraId"] = new SelectList(_context.Carreras, "Id", "Id", curso.CarreraId);
-            ViewData["DocenteId"] = new SelectList(_context.Usuarios, "Id", "Id", curso.DocenteId);
+
+            ViewData["CarreraId"] = new SelectList(_context.Carreras, "Id", "Nombre", curso.CarreraId);
+            ViewData["DocenteId"] = new SelectList(
+                _context.Usuarios.Where(u => u.RolId == 2), "Id", "Nombre", curso.DocenteId);
             return View(curso);
         }
 
-        // GET: Cursoes/Delete/5
+        // GET: /Cursoes/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
+            if (HttpContext.Session.GetString("UsuarioRol") != "Administrador")
+                return RedirectToAction("Index");
+
+            if (id == null) return NotFound();
 
             var curso = await _context.Cursos
                 .Include(c => c.Carrera)
                 .Include(c => c.Docente)
                 .FirstOrDefaultAsync(m => m.Id == id);
-            if (curso == null)
-            {
-                return NotFound();
-            }
 
+            if (curso == null) return NotFound();
             return View(curso);
         }
 
-        // POST: Cursoes/Delete/5
+        // POST: /Cursoes/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (HttpContext.Session.GetString("UsuarioRol") != "Administrador")
+                return RedirectToAction("Index");
+
             var curso = await _context.Cursos.FindAsync(id);
             if (curso != null)
             {
                 _context.Cursos.Remove(curso);
+                await _context.SaveChangesAsync();
+                TempData["Mensaje"] = "Curso eliminado correctamente.";
             }
-
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
